+++ file-service/file-service-api/src/main/java/uk/gov/justice/services/fileservice/api/FileStorer.java
@@ -26,6 +26,4 @@ public interface FileStorer {
      * @param fileId The id of the file to be deleted.
      */
     void delete(final UUID fileId) throws FileServiceException;
-
-    void purgeOldestSoftDeletedFiles() throws FileServiceException;
 }