+++ file-service/file-service-test-utils/src/main/java/uk/gov/justice/services/fileservice/utils/test/FileServiceTestClient.java
@@ -1,5 +1,7 @@
 package uk.gov.justice.services.fileservice.utils.test;
 
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
 import static java.util.UUID.randomUUID;
 import static javax.json.Json.createObjectBuilder;
 
@@ -7,7 +9,9 @@ import uk.gov.justice.services.common.converter.ZonedDateTimes;
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.fileservice.api.FileServiceException;
 import uk.gov.justice.services.fileservice.domain.FileReference;
-import uk.gov.justice.services.fileservice.repository.FileStoreJdbcRepository;
+import uk.gov.justice.services.fileservice.repository.ContentJdbcRepository;
+import uk.gov.justice.services.fileservice.repository.FileContent;
+import uk.gov.justice.services.fileservice.repository.MetadataJdbcRepository;
 
 import java.io.InputStream;
 import java.sql.Connection;
@@ -40,14 +44,16 @@ import javax.json.JsonObject;
  */
 public class FileServiceTestClient {
 
-    private final FileStoreJdbcRepository fileStoreJdbcRepository;
+    private final MetadataJdbcRepository metadataJdbcRepository;
+    private final ContentJdbcRepository contentJdbcRepository;
     private final UtcClock utcClock;
 
     /**
      * Creates a test file service client
      */
     public FileServiceTestClient() {
-        fileStoreJdbcRepository = new FileStoreJdbcRepository();
+        metadataJdbcRepository = new MetadataJdbcRepository();
+        contentJdbcRepository = new ContentJdbcRepository();
         utcClock = new UtcClock();
     }
 
@@ -98,7 +104,9 @@ public class FileServiceTestClient {
                 .add("createdAt", ZonedDateTimes.toString(utcClock.now()))
                 .build();
 
-        fileStoreJdbcRepository.insert(fileId, contentStream, metadata, connection);
+        contentJdbcRepository.insert(fileId, contentStream, connection);
+        metadataJdbcRepository.insert(fileId, metadata, connection);
+
         return fileId;
     }
 
@@ -113,6 +121,23 @@ public class FileServiceTestClient {
      * @throws FileServiceException if retrieving the file fails
      */
     public Optional<FileReference> read(final UUID fileId, final Connection connection) throws FileServiceException {
-        return fileStoreJdbcRepository.findByFileId(fileId, connection);
+
+        final Optional<JsonObject> metadata = metadataJdbcRepository.findByFileId(fileId, connection);
+        final Optional<FileContent> content = contentJdbcRepository.findByFileId(fileId, connection);
+
+        if (metadata.isPresent() && content.isPresent()) {
+
+            final InputStream inputStream = content.get().getContent();
+
+            final FileReference fileReference = new FileReference(
+                    fileId,
+                    metadata.get(),
+                    inputStream,
+                    false);
+
+            return of(fileReference);
+        }
+
+        return empty();
     }
 }